Refinements for temporary velocity limits in multicopter position mode

+++ src/modules/flight_mode_manager/tasks/ManualAltitude/FlightTaskManualAltitude.cpp
@@ -90,12 +90,9 @@ void FlightTaskManualAltitude::_updateConstraintsFromEstimator()
 
 void FlightTaskManualAltitude::_scaleSticks()
 {
-	// Use sticks input with deadzone and exponential curve for vertical velocity
 	float vel_max_z = (_sticks.getPosition()(2) > 0.0f) ? _param_mpc_z_vel_max_dn.get() :
 			  _param_mpc_z_vel_max_up.get();
-
-	vel_max_z = fmin(vel_max_z, _vertical_velocity_limit);
-
+	vel_max_z = fminf(vel_max_z, _vertical_velocity_limit);
 	_velocity_setpoint(2) = vel_max_z * _sticks.getPositionExpo()(2);
 }
